requirements.txt
PyMySQL>=0.8.1
PyYAML>=3.12
